finish new FsLayout.v with proof
new FsLayout proves that its three partitions (inodes, bmaps, and blocks)
are disjoint.

with the help of the previously-proven RefineParts, this also means we
can stack anything on top of the individual partitions, and get the
compiler and proof of the entire FsLayout for free (as long as we
prove refinement for the particular thing we stack).

the proof is, at some level, pretty simple, but ends up long because
a slight detail differs for each of the 3 partitions, and i couldn't
figure out how to automate the "instantiate" tactic.