comment and another issue
the other problem i forgot to mention in the previous commit message is:
we would have to figure out some way to reason about crashes; with this
Execute model, there isn't a clear "step" function at the lowest level.

perhaps we need a hybrid model with two kinds of continuations: either
an invocation of a lower-level operation, or some computation that affects
only the in-memory state of a higher layer?