some sketch of why [donetoken] might be useless
given any statement like [{{pre}} p1 >> p2] and a memory that satisfies
the precondition [pre], it's possible to formally prove the existence
of [donetoken], despite it being an unspecified parameter.

not too surprising, given that Done is the only way to finish a program.
it's not actually possible to get a concrete [donetoken] out of this proof,
because the proof is in [Prop], and [donetoken] is in [Set].

but still, this suggests that perhaps [donetoken] is not so useful after
all, and we should make [Done] actually constructable in coq.  this might
make recovery idempotence easier to prove.