checkpoint some more thoughts on recovery idempotence
now it seems like our goal was bad to start with.  we were trying to
prove something of the form:

  {{ pre }} recover xp rx >> recover xp rx

where [pre] says something very specific about the conditions when
it's OK to invoke [rx], such as, if we were in (ActiveTxn m1 m2) when
we first invoked recover, then [rx] can be called only if the state
matches (NoTransaction m1).

but if we expand out the meaning of this statement, it says that,
starting with precondition [pre], we can run [recover xp rx], and
AT ANY FUTURE POINT, call [recover xp rx] again.  but at some future
point, the continuation [rx] might have changed the state of the disk
(perhaps starting a new transaction, aborting, committing, etc).  thus,
if we THEN run [recover xp rx], the recovery code will roll back to a
DIFFERENT state, which might not match (NoTransaction m1) anymore, so
it wouldn't be legal for it to invoke [rx]..

so perhaps we don't actually want to say the ">> recover xp rx" part.
but then what?