first stab at implementing mutable memory state
new [MPut] and [MGet] opcodes in our language.

one thing i couldn't figure out is a clean way of supporting different
types of variables.  right now, every variable maps to an [mobj]:

  Inductive mobj := MObj (T : Type) (v : T).

but i worry this will be hard to program with, because it'll generate
[existT] terms when destructing or matching on an [mobj]..

next step is to figure out how to write predicates / Hoare statements
about both the persistent disk and the in-memory state.